phase-d(bun_runtime): h2_frame_parser Handlers/Stream — deref event_loop/bun_vm raw ptrs, Err-wrap throw_invalid_arguments, Strong->StrongOptional for js_context/callback, approximate resumable stream iterator, 2-arg constructor for JsClass codegen

+++ src/runtime/image/Image.rs
@@ -870,12 +870,12 @@ pub fn clipboard_change_count(_: &JSGlobalObject, _: &CallFrame) -> JsResult<JSV
 impl Image {
     #[bun_jsc::host_fn(getter)]
     pub fn get_width(&self, _: &JSGlobalObject) -> JSValue {
-        JSValue::js_number(self.last_width)
+        JSValue::js_number(f64::from(self.last_width))
     }
 
     #[bun_jsc::host_fn(getter)]
     pub fn get_height(&self, _: &JSGlobalObject) -> JSValue {
-        JSValue::js_number(self.last_height)
+        JSValue::js_number(f64::from(self.last_height))
     }
 }
